ai_recruitment_system/requirements.txt
# AI Recruitment System Requirements
# Core Streamlit and AI dependencies
streamlit>=1.28.0
langchain>=0.3.26
langchain-core>=0.3.69
langchain-community>=0.3.27
langchain-openai>=0.3.28
langgraph>=0.5.3
openai>=1.97.0

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11

# Environment management
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Additional utilities
requests>=2.31.0
typing-extensions>=4.8.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
